requirements.txt
click
python-dotenv
terminaltables

requests
